+++ src/trx/game/lara/mesh.h
@@ -10,5 +10,4 @@ void Lara_Mesh_SwapSingle(LARA_MESH mesh, OBJECT_ID obj_id);
 void Lara_Mesh_SwapAll(OBJECT_ID obj_id);
 void Lara_Mesh_Set(LARA_MESH mesh, OBJECT_MESH *mesh_ptr);
 OBJECT_MESH *Lara_Mesh_Get(LARA_MESH mesh);
-RGBA_F Lara_GetMeshTint(GAME_VECTOR pos);
 int32_t Lara_GetMeshIndex(const ITEM *item, int32_t mesh_idx);